Add files via upload
In this project, we will look at how you can emulate a player throwing a set of dice. Then we 
will look at how you can store these dice in a pile the player wants to keep. you write a game where 
you throw dice to determine the outcome, this is one way to store the results for later use.