Add files via upload
A row of dashes represents the word to guess. If the player guesses a letter in the word, the 
script writes it in all its correct positions. The player has infinite turns to guess the word depends upon the level he chose. You can easily 
customize the game by changing the variables.